Start some refactoring
* remove stuff we're not using yet/anymore but keep for later
* move file path var setting to module to get out of the way
* start dissecting some methods to prepare for bigger refactor
